% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic-form.R
\name{CalcQuadFormP}
\alias{CalcQuadFormP}
\title{Calculate P-value for Quadratic Form}
\usage{
CalcQuadFormP(a, test_stat, v, eps = 1e-08, method = "davies")
}
\arguments{
\item{a}{Central matrix of the quadratic form.}

\item{test_stat}{Test statistic.}

\item{v}{Variance of the random vector.}

\item{eps}{Minimum eigenvalue magnitude.}

\item{method}{"davies" or "liu". If "davies", "liu" will provide a fallback.}
}
\value{
Numeric p-value.
}
\description{
Calculate the p-value for a statistic \deqn{T = y^{\top}Ay}, where
\deqn{y ~ N(0, v)} and A is any symmetric positive-definite matrix.
}
