% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Corr}
\alias{Corr}
\title{Marginal correlations}
\usage{
Corr(y, x)
}
\arguments{
\item{y}{(n x 1) Numeric vector.}

\item{x}{(n x p) Numeric matrix.}
}
\value{
List containing the following:
\itemize{
\item{r: Correlations between y and the columns of x.}
\item{R: Correlations among the columns of x.}
}
}
\description{
Fits the standard OLS model.
}
