% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Expected}
\alias{Expected}
\title{Expected values}
\usage{
Expected(beta_g, mu_g, var_g, beta_e, mu_e, var_e, beta_h, var_resid)
}
\arguments{
\item{beta_g}{Joint coefficient for G.}

\item{mu_g}{Mean of G.}

\item{var_g}{Variance of G.}

\item{beta_e}{Joint coefficient for E.}

\item{mu_e}{Mean of E.}

\item{var_e}{Variance of E.}

\item{beta_h}{Joint coefficient for H = G x E.}

\item{var_resid}{Residual variance.}
}
\value{
List of named expected values.
}
\description{
Calculates expected values under the generative model:
\deqn{Y = G\beta_{G} + E\beta_{E} + H\beta_{H} + \epsilon}
}
