% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FitOLS}
\alias{FitOLS}
\title{Ordinary Least Squares}
\usage{
FitOLS(y, X)
}
\arguments{
\item{y}{n x 1 Numeric vector.}

\item{X}{n x p Numeric matrix.}
}
\value{
List containing the following:
\item{Beta}{Regression coefficient.}
\item{V}{Outcome variance.}
\item{Ibb}{Information matrix for beta.}
\item{Resid}{Outcome residuals.}
}
\description{
Fits the standard OLS model.
}
