% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generation.R
\name{GenCovar}
\alias{GenCovar}
\title{Generate Covariates}
\usage{
GenCovar(n, maf)
}
\arguments{
\item{n}{Sample size.}

\item{maf}{Minor allele frequency.}
}
\value{
n x 3 covariate matrix.
}
\description{
Generate Covariates
}
