% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{PrattIFTest}
\alias{PrattIFTest}
\title{Pratt Influence Function Test}
\usage{
PrattIFTest(y, g, e)
}
\arguments{
\item{y}{(n x 1) numeric vector. Phenotype.}

\item{g}{(n x 1) numeric vector. Genotype.}

\item{e}{(n x 1) numeric vector. Environment.}
}
\value{
Data.frame of results.
}
\description{
Computes Wald tests for the Pratt indices corresponding to G, E, and
H = G * E with inference via the empirical influence function.
}
