% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrattIndex-package.R
\docType{package}
\name{PrattIndex-package}
\alias{PrattIndex-package}
\alias{_PACKAGE}
\title{PrattIndex: Inference on the Pratt Index}
\description{
Inference on the Pratt Index.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zrmacc/PrattIndex}
}

}
\author{
\strong{Maintainer}: Zachary McCaw \email{zmccaw@alumni.harvard.edu} (\href{https://orcid.org/0000-0002-2006-9828}{ORCID})

}
\keyword{internal}
