% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PrattIndex}
\alias{PrattIndex}
\title{Pratt Index}
\usage{
PrattIndex(y, g, e)
}
\arguments{
\item{y}{(n x 1) Numeric vector.}

\item{g}{(n x 1) Genotype vector.}

\item{e}{(n x 1) Environment vector.}
}
\value{
List containing the following:
\itemize{
\item{beta: (4 x 1) vector of regression coefficients (intercept, G, E, H).}
\item{kappa: (3 x 1) vector of Pratt indices (G, E, H).}
\item{grad_kappa: (3 x 3) matrix of Pratt index gradients with respect to beta, column-wise.}
\item{rho: (3 x 1) vector of correlations (G, E, H).}
\item{var_beta: (4 x 4) covariance matrix of beta.}
\item{var_kappa: (3 x 1) vector of Pratt index variances (G, E, H).}
\item{var_resid: Residual variance from regression of y ~ (intercept, G, E, H).}
}
}
\description{
Calculate the Pratt index and intermediates.
}
