% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PrattIndexCpp}
\alias{PrattIndexCpp}
\title{Pratt Index}
\usage{
PrattIndexCpp(y, g, e)
}
\arguments{
\item{y}{(n x 1) numeric}

\item{g}{(n x 1) numeric}

\item{e}{(n x 1) numeric}
}
\value{
List containing:
\itemize{
\item beta: Regression coefficients (b_G, b_E, b_H) from the joint model.
\item beta_null: Regression coefficient (b_G, b_E) from the model that omits H.
\item cov_xx: Covariance matrix of (G, E, H).
\item kappa: Pratt index
\item grad_kappa: Gradient of kappa with respect to beta.
\item var_beta: Covariance matrix of beta.
\item var_beta_null: covariance matrix of beta from the model that omits H.
\item var_kappa: Variance of Pratt indices.
\item var_resid: Residual variance from the model that includes H.
\item var_resid_null: Residual variance from the model the omits H.
}
}
\description{
Pratt Index
}
