% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PrattInfluenceCpp}
\alias{PrattInfluenceCpp}
\title{Pratt Index Influence Function}
\usage{
PrattInfluenceCpp(y, g, e)
}
\arguments{
\item{y}{(n x 1) numeric}

\item{g}{(n x 1) numeric}

\item{e}{(n x 1) numeric}
}
\value{
(n x 3) numeric matrix psi, where psi.row(i) is IF_kappa for obs i.
}
\description{
Computes the plug-in influence function for the Pratt index components
corresponding to X = (G, E, H) with H = G * E.
}
