% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{ScoreTest}
\alias{ScoreTest}
\title{Score Test}
\usage{
ScoreTest(beta_null, cov_xx, kappa_h, n, var_resid_null, var_y)
}
\arguments{
\item{beta_null}{Estimate of (b_G, b_E) from the model that omits H.}

\item{cov_xx}{Covariance matrix of (G, E, H).}

\item{kappa_h}{Pratt index for H.}

\item{n}{Sample size.}

\item{var_resid_null}{Estimate of residual variance from the model that omits H.}

\item{var_y}{Marginal variance of Y.}
}
\description{
Score Test
}
