% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic-form.R
\name{CalcQuadFormP}
\alias{CalcQuadFormP}
\title{Calculate P-value for Quadratic Form}
\usage{
CalcQuadFormP(cov_z, kernel, test_stat, eps = 1e-08, method = "davies")
}
\arguments{
\item{cov_z}{Covariance matrix of the random vector.}

\item{kernel}{Central matrix of the quadratic form.}

\item{test_stat}{Test statistic.}

\item{eps}{Minimum eigenvalue magnitude.}

\item{method}{"davies" or "liu". If "davies", "liu" will provide a fallback.}
}
\value{
Numeric p-value.
}
\description{
Calculate the p-value for a statistic \deqn{T = z^{\top}Az}, where
\deqn{z ~ N(0, cov_z)} and A is any symmetric positive-definite matrix.
}
