% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FitOLS}
\alias{FitOLS}
\title{Ordinary Least Squares}
\usage{
FitOLS(y, X)
}
\arguments{
\item{y}{n x 1 numeric vector.}

\item{X}{n x p numeric matrix.}
}
\value{
List with \code{beta}, \code{resid_var}, \code{var_beta}.
}
\description{
Ordinary Least Squares
}
