% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generation.R
\name{GenData}
\alias{GenData}
\title{Generate Data}
\usage{
GenData(
  beta_g = 0.1,
  beta_e = 0.1,
  beta_h = 0,
  maf = 0.25,
  mu_g = NULL,
  mu_e = 0,
  n = 1000,
  n_e = NULL,
  p_e = NULL,
  type_e = "normal",
  type_g = "binom",
  var_e = 1,
  var_exp = 0.25,
  var_g = NULL,
  var_resid = NULL
)
}
\arguments{
\item{beta_g}{Genetic beta.}

\item{beta_e}{Environment beta.}

\item{beta_h}{Interaction beta.}

\item{maf}{Minor allele frequency for G, if type_g = "binom".}

\item{mu_g}{Mean of G, if type_g = "normal".}

\item{mu_e}{Mean of E, if type_e = "normal".}

\item{n}{Sample size.}

\item{n_e}{Number of trials for E, if type_e = "binom".}

\item{p_e}{Probability of success for E, if type_e = "binom".}

\item{type_e}{Either "binom" or "normal".}

\item{type_g}{Either "binom" or "normal".}

\item{var_e}{Variance of E, if type_e = "normal".}

\item{var_exp}{Proportion of variance in Y explained by (G, E, H).}

\item{var_g}{Variance of G, if type_g = "normal".}

\item{var_resid}{Residual variance. Set to null to specify \code{var_exp} instead.}
}
\description{
Simulates a data set containing the phenotype, genotype, environment, and
interaction. Genotype can be either binomial or normal. If binomial, specify the
minor allele frequency. If normal, specify mu_g and var_g. Likewise, environment
can be binomial or normal. If binomial, specify n_e and p_e. If normal, specify
mu_e and var_e. The residual phenotypic variance can either be specified directly,
or it can be inferred by specifying the variance explained by (G, E, H).
}
