% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generation.R
\name{GenData}
\alias{GenData}
\title{Generate Data}
\usage{
GenData(n = 1000, beta_g = 0.1, beta_e = 0.1, beta_h = 0, maf = 0.25)
}
\arguments{
\item{n}{Sample size.}

\item{beta_g}{Genetic beta.}

\item{beta_e}{Environment beta.}

\item{beta_h}{Interaction beta.}

\item{maf}{Genetic minor allele frequency.}
}
\description{
Generate Data
}
