% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generation.R
\name{GenPheno}
\alias{GenPheno}
\title{Generate Phenotype}
\usage{
GenPheno(beta, x)
}
\arguments{
\item{beta}{3 x 1 beta vector for the joint model.}

\item{x}{n x 3 covariate data.frame.}
}
\value{
n x 1 phenotype vector.
}
\description{
Uses the centered g and e to form the interaction.
}
