% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generation.R
\name{GenPheno}
\alias{GenPheno}
\title{Generate Phenotype}
\usage{
GenPheno(beta, x, var_exp, var_resid = NULL)
}
\arguments{
\item{beta}{3 x 1 beta vector for the joint model.}

\item{x}{n x 3 covariate data.frame.}

\item{var_exp}{Proportion of variance explained by main effects.}

\item{var_resid}{Residual variance explained. Set to null to specify \code{var_exp} instead.}
}
\value{
n x 1 phenotype vector.
}
\description{
Uses the centered g and e to form the interaction.
}
