% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MatrixSqrt}
\alias{MatrixSqrt}
\title{Matrix Square Root}
\usage{
MatrixSqrt(S, eps = 1e-08)
}
\arguments{
\item{S}{(p x p) numeric matrix.}

\item{eps}{Scalar minimum eigenvalue.}
}
\description{
Matrix Square Root
}
