% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrattIndex-package.R
\docType{package}
\name{PrattIndex-package}
\alias{PrattIndex-package}
\title{PrattIndex: Inference on the Pratt Index}
\description{
Inference on the Pratt Index.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zrmacc/PrattIndex}
}

}
\keyword{internal}
