% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{PrattIndex}
\alias{PrattIndex}
\title{Pratt Index}
\usage{
PrattIndex(y, g, e)
}
\arguments{
\item{y}{Phenotype.}

\item{g}{Genotype.}

\item{e}{Environment.}
}
\value{
List of results.
}
\description{
Pratt Index
}
