% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PrattIndex}
\alias{PrattIndex}
\title{Pratt Index}
\usage{
PrattIndex(y, x, standard = TRUE)
}
\arguments{
\item{y}{(n x 1) Numeric vector.}

\item{x}{(n x p) Numeric matrix.}

\item{standard}{Standardize y and x?}
}
\value{
List containing the following:
\itemize{
\item{beta: Regression coefficients.}
\item{r: Correlations between y and the columns of x.}
\item{R: Correlations among the columns of x.}
\item{pratt: Pratt indices for the columns of x.}
\item{var_exp: Variance in y explained by x.}
}
}
\description{
Calculate the Pratt index and intermediates.
}
