% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PrattIndex}
\alias{PrattIndex}
\title{Pratt Index}
\usage{
PrattIndex(y, g, e)
}
\arguments{
\item{y}{(n x 1) Numeric vector.}

\item{g}{(n x 1) Genotype vector.}

\item{e}{(n x 1) Environment vector.}
}
\value{
List containing the following:
\itemize{
\item{beta: (3 x 1) vector of regression coefficient.}
\item{pratt: Scalar Pratt index of the interaction term.}
\item{sigma2: Scalar residual variance from the joint model.}
\item{var_beta: (3 x 3) sampling variance of the regression coefficients.}
\item{var_exp: Scalar variance in y explained by x.}
\item{var_pratt: Scalar sampling variance of the pratt index.}
}
}
\description{
Calculate the Pratt index and intermediates.
}
