% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{PrattTest}
\alias{PrattTest}
\title{Pratt Test}
\usage{
PrattTest(y, g, e)
}
\arguments{
\item{y}{Phenotype.}

\item{g}{Genotype.}

\item{e}{Environment.}
}
\value{
Data.frame of results.
}
\description{
Performs Wald tests for the Pratt indices of G, E, and H = G*E using the
influence-function-based variance estimator.
}
