% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pratt-test.R
\name{PrattTest}
\alias{PrattTest}
\title{Pratt Test}
\usage{
PrattTest(y, g, e, alpha = 0.05)
}
\arguments{
\item{y}{Phenotype.}

\item{g}{Genotype vector.}

\item{e}{Environmental exposure.}

\item{alpha}{Alpha for confidence intervals.}
}
\value{
Data.frame.
}
\description{
Pratt Test
}
