% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pratt-test.R
\name{PrattTest}
\alias{PrattTest}
\title{Pratt Test}
\usage{
PrattTest(g, e, y, standardize = TRUE)
}
\arguments{
\item{g}{Genotype vector.}

\item{e}{Environmental exposure.}

\item{y}{Phenotype.}

\item{standardize}{Standardize?}
}
\value{
Data.frame.
}
\description{
Pratt Test
}
