% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{PrattTest}
\alias{PrattTest}
\title{Pratt Test}
\usage{
PrattTest(y, g, e, use_score_test = TRUE)
}
\arguments{
\item{y}{Phenotype.}

\item{g}{Genotype.}

\item{e}{Environment.}

\item{use_score_test}{Logical.}
}
\value{
Data.frame of results.
}
\description{
Performs the Pratt test for the interaction between G and E.
}
